inspection/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
packages = [ { include = "lib" } ]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.13"}
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = ["tests/*"]

[tool.pytest.ini_options]
log_cli = false
log_cli_level = "INFO"

[tool.ruff]
# Allow lines to be as long as 100 characters.
line-length = 100
exclude = ["tests"]
